requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==3.0.2
WTForms==3.1.0
email-validator==2.1.1
gunicorn==21.2.0
numpy==1.26.4
openai==1.54.0
pandas==2.2.2
plotly==5.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
yfinance==0.2.38
scipy==1.12.0
SQLAlchemy==2.0.28
alembic==1.13.1
psycopg[binary]
Pillow==10.2.0
beautifulsoup4==4.12.2
Flask-Dance==7.1.0
oauthlib==3.2.2
schedule==1.2.0
sendgrid==6.11.0
stripe==7.11.0
pydantic==2.5.0
jinja2==3.1.2
